src/processing/requirements_spark.txt
pika 
py4j 
requests 
pandas 
plotly 
sqlalchemy
psycopg2-binary 
openpyxl 
tqdm 